requirements/dev.txt
# Everything the developer needs in addition to the production requirements
-r prod.txt

# Debug toolbar
django-debug-toolbar==1.9.1

# Testing
factory-boy==2.10.0
selenium==3.9.0
coveralls~=1.2.0
chromedriver-binary~=2.35.1

# Lint and code style
flake8==3.5.0
flake8-blind-except==0.1.1
flake8-isort==2.3
flake8-debugger==3.1.0
flake8-docstrings==1.3.0
flake8-quotes==0.13.0
pycodestyle~=2.3.1
pep8-naming==0.5.0
pylint~=1.8.2
pylint-django~=0.9.0
yapf~=0.20.2

rope~=0.10.7